+++ DotnetCoreNLayer.Core/Repositories/IProductRepository.cs
@@ -8,6 +8,6 @@ namespace DotnetCoreNLayer.Core.Repositories
 {
     public interface IProductRepository : IRepository<Product>
     {
-        Task<Product> GetWithCategoryByIdAsync(int productId);
+        Task<Product> GetWithCategoryByIdAsync(long productId);
     }
 }
